+++ src/cpu/cpu_constants.js
@@ -60,3 +60,5 @@ define('ORCC', 44);
 define('READEOR', 45);
 define('SHIFTLEFT', 46);
 define('SHIFTRIGHT', 47);
+define('ROTATELEFT', 48);
+define('ROTATERIGHT', 49);